+++ app/Services/Shaarli/Shaarli.php
@@ -73,80 +73,7 @@ class Shaarli
 
     public function getSettingsConfig(): array
     {
-        return [
-            'name' => [
-                'default' => config('app.name'),
-                'rules' => ['required', 'min:2', 'max:100']
-            ],
-            'locale' => [
-                'default' => config('app.locale'),
-                'rules' => ['required', 'in:fr,en,de']
-            ],
-            'is_private' => [
-                'default' => false,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'is_dark' => [
-                'default' => false,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'home_show_tags' => [
-                'default' => true,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'home_show_chests' => [
-                'default' => true,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'compact_cardslist' => [
-                'default' => false,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'columns_count' => [
-                'default' => 3,
-                'rules' => ['required', 'numeric', 'min:1', 'max:4']
-            ],
-            'custom_background' => [
-                'default' => null,
-                'rules' => ['nullable', 'url']
-            ],
-            'custom_background_encoded' => [
-                'default' => null,
-                'rules' => ['nullable']
-            ],
-            'private_archive' => [
-                'default' => false,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'secure_login' => [
-                'default' => false,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'secure_code_expires' => [
-                'default' => 30,
-                'rules' => ['required', 'numeric', 'min:5', 'max:300']
-            ],
-            'secure_code_length' => [
-                'default' => 8,
-                'rules' => ['required', 'numeric', 'min:4', 'max:12']
-            ],
-            'link_archive_pdf' => [
-                'default' => true,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'link_archive_media' => [
-                'default' => true,
-                'rules' => ['nullable', 'in:on,off']
-            ],
-            'node_bin' => [
-                'default' => '/usr/bin/node',
-                'rules' => ['required']
-            ],
-            'youtube_dl_bin' => [
-                'default' => '/usr/bin/youtube-dl',
-                'rules' => ['required']
-            ],
-        ];
+        return $this->app['config']->get('shaarli.settings');
     }
 
     public function getSettings(): array
@@ -165,7 +92,7 @@ class Shaarli
         );
     }
 
-    public function setSettings(Collection $settings): void
+    public function setSettings(Collection $settings): self
     {
         foreach ($this->getSettingsConfig() as $key => $item) {
             if (is_bool($item['default'])) {
@@ -198,6 +125,22 @@ class Shaarli
 
             $this->settings->put($key, $settings->get($key, $item['default']));
         }
+
+        return $this;
+    }
+
+    public function cleanSettings(): self
+    {
+        $defaults = $this->getSettingsConfig();
+
+        foreach ($this->settings as $key => $value)
+        {
+            if (false === array_key_exists($key, $defaults)) {
+                $this->settings->forget($key);
+            }
+        }
+
+        return $this;
     }
 
     public function __call($name, $arguments)